Don't call aref on undefined variable
Calling `(aref ansi-term-color-vector 0)` when `ansi-term-color-vector`
isn't already defined results an exception, so check to make sure it
exists before investigating it.

I defaulted to setting the variable if it doesn't already exist so it's
loaded if someone calls ansi-term later, presuming that they'll
overwrite this variable if they don't want to use farmhouse theme for it.